Additions and Fixes
Added more sounds

Losing lives sound
Stock panel open and close sound
Changed Lightning Tower shooting sound

Fixed towers not selling AGAIN. It was also the script that wasn't working well. It works fine now.

Added renderer color changes when tower is not on/not on buildable areas. Green for buildable and red for not buildable